Login Page
PHP MySQL Login System
In this topic, we will learn how to create a PHP MySQL Login System with the help of PHP and MySQL database. There are few steps given for creating a login system with MySQL database.

Before creating the login system first, we need to know about the pre-requisites to create the login module.

Requirements
We should have knowledge of HTML, CSS, PHP and MySQL for creating the login system.
Text Editor - For writing the code. We can use any text editor such as Notepad, Notepad++, Dreamweaver, etc.
XAMPP - XAMPP is a cross-platform software, which stands for Cross-platform(X) Apache server (A), MySQL (M), PHP (P), Perl (P). XAMPP is a complete software package, so, we don't need to install all these separately.
Environment Setup
Now, we need to start the webserver and create the files for the login system. There are few steps that are given below to setup the environment.

Open the XAMPP Control Panel.
Start the Apache server by clicking on the Start button.
Start the MySQL by clicking on the Start button.
Create all the files needed for login.
Create login table in the database using phpMyAdmin in XAMPP.
Now, we will create four files here for the login system.

index.html - This file is created for the GUI view of the login page and empty field validation.
style.css - This file is created for the attractive view of the login form.
connection.php - Connection file contains the connection code for database connectivity.
authentication.php - This file validates the form data with the database which is submitted by the user.
Save all these files in the htdocs folder inside the Xampp installation folder. The detailed description and code for these files are discussed below.